Moved helper js files to /lib
