+++ lib/prediction-ucf-garage.js
